Fix copy constructor and assignment operator of PhaseSpace
In the process, the swap method was optimized.
Also, PhaseSpace::_data now is a boost::multi_array
instead of a Array::array3. This would not have been needed
but was done to be able to debug well documented code.
(Array.h will be replaced anyways, as we rely on custom
additions to allow more stl-like use. Those, however,
were not accepted upstream.)