Fix widening compare in loops (#114)
As this is a rather unimportant fix, it is not needed for v1.1 and filed against dev.

Remark: It would be thinkable that the value of _ip exceeds the limit of uint_fast8_t, causing an infinite loop. However, it is known (but not guaranteed by the file type) that _ip <= 4.